Don't block the build for translation errors (#19)
This allows the translations lint/test step to fail without failing the whole job and thus blocking the build. It also logs some summary info for the workflow so it's possible to keep an eye on the status of translation issues, and helps keep `src/i18n/en.json` up to date (for example, if you committed from the GitHub web UI, where pre-commit hooks don't run).